refactor: refactor configuration management and simplify provider creation
